Added controller to run on Arduino Mega
Introduced file "SUITCEYES_Controller_1_0_Mega_Master.ino", so that an Arduino Mega can be used as controller. As provided, it can run up to 30 channels (up to 12 with PWM). Due to the Arduino serial buffer constraint, a Packet Size of 1 is used, so intensities are specified as either 0 (off) or 1 (on), and durations are specified in multiples of 100ms, from 0 (0ms) to 9 (900ms). It is possible to raise the packet size if fewer channels are required -  a packet size of 2 will permit up to 20 channels. A packet size of 3 will permit 12 channels, so that all 12 PWM channels can be used.